application/requirements.txt
# Production
pyodbc
flask
flask_bootstrap4

# Development
pip
flake8
isort
pylama
pylint
ipython
autopep8
